Add converters and insert Products
ListToString and StringToList converters were added to hanlde complex data within Room.

Products are now properly created and updated. Previous existing data is shown on init.

Packaging and Ingredients association with Products has been implemented but not yet tested.

+++ app/src/main/java/com/example/atry/database/IngredientDao.kt
@@ -18,7 +18,7 @@ interface IngredientDao {
     fun update(ingredient: Ingredient)
 
     @Query("SELECT * from ingredient_table WHERE id = :key")
-    fun get(key: Long): Ingredient?
+    fun get(key: Int): Ingredient?
 
     @Query("DELETE FROM ingredient_table")
     fun clear()
